Support for Android 16

+++ app/src/main/java/arunkbabu90/popmovies/ui/fragment/FavouritesFragment.kt
@@ -18,6 +18,7 @@ import arunkbabu90.popmovies.Constants
 import arunkbabu90.popmovies.R
 import arunkbabu90.popmovies.data.model.Favourite
 import arunkbabu90.popmovies.data.model.FavouritesLiveData
+import arunkbabu90.popmovies.databinding.FragmentFavouritesBinding
 import arunkbabu90.popmovies.ui.activity.MovieDetailsActivity
 import arunkbabu90.popmovies.ui.adapter.FavouritesAdapter
 import arunkbabu90.popmovies.ui.viewmodel.FavouritesViewModel
@@ -29,10 +30,11 @@ import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.SetOptions
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import kotlinx.android.synthetic.main.fragment_favourites.*
-import kotlinx.android.synthetic.main.item_favourites.*
 
 class FavouritesFragment : Fragment() {
+    private var _binding: FragmentFavouritesBinding? = null
+    private val binding get() = _binding!!
+
     private lateinit var adapter: FavouritesAdapter
 
     private val db: FirebaseFirestore = Firebase.firestore
@@ -41,9 +43,9 @@ class FavouritesFragment : Fragment() {
     private var isScrolling = false
     private var favouritesLiveData: FavouritesLiveData? = null
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favourites, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding = FragmentFavouritesBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -53,16 +55,16 @@ class FavouritesFragment : Fragment() {
         val path = "${Constants.COLLECTION_USERS}/${user.uid}/${Constants.COLLECTION_FAVOURITES}"
 
         adapter = FavouritesAdapter(favouriteMovies) { favouriteMovie ->
-            if (favouriteMovie != null) onFavouriteClick(favouriteMovie)
+            if (favouriteMovie != null) onFavouriteClick(favouriteMovie, null)
         }
-        tv_fav_err.visibility = if (favouriteMovies.isNullOrEmpty()) View.VISIBLE else View.GONE
+        binding.tvFavErr.visibility = if (favouriteMovies.isNullOrEmpty()) View.VISIBLE else View.GONE
 
         val lm = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
-        rv_favourites.setHasFixedSize(true)
-        rv_favourites.layoutManager = lm
-        rv_favourites.adapter = adapter
+        binding.rvFavourites.setHasFixedSize(true)
+        binding.rvFavourites.layoutManager = lm
+        binding.rvFavourites.adapter = adapter
 
-        rv_favourites.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        binding.rvFavourites.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
                 super.onScrollStateChanged(recyclerView, newState)
                 if (newState == AbsListView.OnScrollListener.SCROLL_STATE_TOUCH_SCROLL)
@@ -77,7 +79,6 @@ class FavouritesFragment : Fragment() {
                 val totalMovieCount = lm.itemCount
 
                 if (isScrolling && (firstVisibleMoviePos + visibleMovieCount == totalMovieCount)) {
-                    // End Of Page; So load more movies if available
                     isScrolling = false
                     getMovies()
                 }
@@ -88,54 +89,54 @@ class FavouritesFragment : Fragment() {
 
         ItemTouchHelper(object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
             override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder,
-                                target: RecyclerView.ViewHolder) = false
+                                target: RecyclerView.ViewHolder): Boolean = false
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                // Delete the Favourite Movie from the list (database)
-                val holder = viewHolder as FavouritesAdapter.FavouritesViewHolder
-                val movie: Favourite = holder.movie ?: return
+                val favAdapterViewHolder = viewHolder as FavouritesAdapter.FavouritesViewHolder
+                // Accessing 'movie' property requires it to be public in FavouritesAdapter.FavouritesViewHolder
+                // For now, let's assume you'll get it from favouriteMovies list using adapterPosition
+                val position = favAdapterViewHolder.absoluteAdapterPosition
+                if (position != RecyclerView.NO_POSITION) {
+                    val movie: Favourite = favouriteMovies[position]
 
-                db.collection(path)
-                    .document(movie.movieId)
-                    .delete()
-                    .addOnSuccessListener {
-                        // Undo Action
-                        val snackbar = Snackbar.make(favourites_fragment_layout, "${movie.title} Removed", Snackbar.LENGTH_LONG)
-                            .setAction(getString(R.string.undo)) {
-                                val favMovie = hashMapOf(
-                                    Constants.FIELD_TITLE to movie.title,
-                                    Constants.FIELD_POSTER_PATH to movie.posterPath,
-                                    Constants.FIELD_BACKDROP_PATH to movie.backdropPath,
-                                    Constants.FIELD_OVERVIEW to movie.overview,
-                                    Constants.FIELD_RELEASE_DATE to movie.releaseDate,
-                                    Constants.FIELD_RATING to movie.rating,
-                                    Constants.FIELD_TIMESTAMP to Timestamp.now())
+                    db.collection(path)
+                        .document(movie.movieId)
+                        .delete()
+                        .addOnSuccessListener {
+                            val snackbar = Snackbar.make(binding.favouritesFragmentLayout, "${movie.title} Removed", Snackbar.LENGTH_LONG)
+                                .setAction(getString(R.string.undo)) {
+                                    val favMovieMap = hashMapOf(
+                                        Constants.FIELD_TITLE to movie.title,
+                                        Constants.FIELD_POSTER_PATH to movie.posterPath,
+                                        Constants.FIELD_BACKDROP_PATH to movie.backdropPath,
+                                        Constants.FIELD_OVERVIEW to movie.overview,
+                                        Constants.FIELD_RELEASE_DATE to movie.releaseDate,
+                                        Constants.FIELD_RATING to movie.rating,
+                                        Constants.FIELD_TIMESTAMP to Timestamp.now())
 
-                                db.collection(path).document(movie.movieId)
-                                    .set(favMovie, SetOptions.merge())
-                            }
-                        snackbar.setActionTextColor(ContextCompat.getColor(requireContext(), R.color.colorGreen))
-                        snackbar.show()
-                    }
+                                    db.collection(path).document(movie.movieId)
+                                        .set(favMovieMap, SetOptions.merge())
+                                }
+                            snackbar.setActionTextColor(ContextCompat.getColor(requireContext(), R.color.colorGreen))
+                            snackbar.show()
+                        }
+                }
             }
-        }).attachToRecyclerView(rv_favourites)
+        }).attachToRecyclerView(binding.rvFavourites)
     }
 
-    /**
-     * Called when a Favourite Movie is clicked
-     * @param favMovie Favourite  The favourite movie that is clicked
-     */
-    private fun onFavouriteClick(favMovie: Favourite) {
+    private fun onFavouriteClick(favMovie: Favourite, posterView: View?) {
         val intent = Intent(activity, MovieDetailsActivity::class.java)
-        val posterView = iv_fav_poster
-        val transitionOptions: ActivityOptionsCompat? =
-            activity?.let { activity ->
+
+        val options: ActivityOptionsCompat? = activity?.let { activity ->
+            posterView?.let {
                 ActivityOptionsCompat.makeSceneTransitionAnimation(
                     activity,
-                    posterView,
-                    ViewCompat.getTransitionName(posterView) ?: "null"
+                    it,
+                    ViewCompat.getTransitionName(it) ?: "nullTransNameFav"
                 )
             }
+        }
 
         intent.putExtra(MovieDetailsActivity.KEY_MOVIE_ID_EXTRA, favMovie.movieId.toInt())
         intent.putExtra(MovieDetailsActivity.KEY_POSTER_PATH_EXTRA, favMovie.posterPath)
@@ -145,10 +146,7 @@ class FavouritesFragment : Fragment() {
         intent.putExtra(MovieDetailsActivity.KEY_RELEASE_DATE_EXTRA, favMovie.releaseDate)
         intent.putExtra(MovieDetailsActivity.KEY_TITLE_EXTRA, favMovie.title)
 
-        if (transitionOptions != null)
-            startActivity(intent, transitionOptions.toBundle())
-        else
-            startActivity(intent)
+        startActivity(intent, options?.toBundle())
     }
 
     private fun getViewModel() = ViewModelProvider(this).get(FavouritesViewModel::class.java)
@@ -160,13 +158,11 @@ class FavouritesFragment : Fragment() {
         favouritesLiveData?.observe(viewLifecycleOwner) { operation ->
             when (operation.type) {
                 R.string.add_operation -> {
-                    // Add
                     val addedMovie = operation.favMovie
                     if (!favouriteMovies.contains(addedMovie))
                         favouriteMovies.add(addedMovie)
                 }
                 R.string.modify_operation -> {
-                    // Modify
                     val modifiedMovie = operation.favMovie
                     for ((i, currentMovie) in favouriteMovies.withIndex())
                         if (currentMovie.movieId == modifiedMovie.movieId) {
@@ -175,7 +171,6 @@ class FavouritesFragment : Fragment() {
                         }
                 }
                 R.string.remove_operation -> {
-                    // Remove
                     val removedMovie = operation.favMovie
                     if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
                         favouriteMovies.removeIf { it.movieId == removedMovie.movieId }
@@ -189,13 +184,14 @@ class FavouritesFragment : Fragment() {
                     }
                 }
             }
-            tv_fav_err.visibility = if (favouriteMovies.isNullOrEmpty()) View.VISIBLE else View.GONE
+            binding.tvFavErr.visibility = if (favouriteMovies.isNullOrEmpty()) View.VISIBLE else View.GONE
             adapter.notifyDataSetChanged()
         }
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
+    override fun onDestroyView() {
+        super.onDestroyView()
         favouritesLiveData?.removeEventListener()
+        _binding = null
     }
 }
